Create Dijkstra's SPF
Lines 1 - 4 :
SetLoops gives a single number of a set at random (faster then set -> list at the cost of randomizing value obtained)  ; for Dijkstra's algo the cost doesn't matter as order of nodes is inconsequential.

Lines 6 - 10 :
Adjacency List used to represent the graph (As of now values are bidirectional although it is possible to change the values to be unidirectional)

Lines 13 onward :
The algorithm itself. Lines 49 to 51 is just a stylistic choice of representation of the results.